Add reference doc to left tab panel
[#174542553]
