+++ src/components/document/document-or-browser.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { useAppConfigStore } from "../../hooks/use-stores";
 import { ELeftTab, LeftTabSpec } from "../../models/view/left-tabs";
-import { DocumentTabContent } from "../navigation/document-tab-content";
+import { DocumentTabPanel } from "../navigation/document-tab-panel";
 import { EditableDocumentContent, IProps as IEditableDocumentContentProps } from "./editable-document-content";
 
 interface IDocumentOrBrowserProps extends IEditableDocumentContentProps {
@@ -11,7 +11,7 @@ interface IDocumentOrBrowserProps extends IEditableDocumentContentProps {
 export const DocumentOrBrowser: React.FC<IDocumentOrBrowserProps> = props => {
   const { showBrowser, tabSpec, ...others } = props;
   return showBrowser && tabSpec
-          ? <DocumentTabContent tabSpec={tabSpec} />
+          ? <DocumentTabPanel tabSpec={tabSpec} />
           : <EditableDocumentContent {...others} />;
 };
