+++ src/components/navigation/problem-panel.sass
@@ -3,7 +3,7 @@
 .problem-panel
   height: 100%
   .section
-    height: calc(100vh - (#{$header-height} + #{$workspace-content-margin} * 2 + #{$left-tab-height} * 2))
+    height: calc(100vh - (#{$header-height} + #{$workspace-content-margin} * 2 + #{$left-tab-height} * 2 + #{$tab-section-border-width} * 2))
     .section-header
       height: $problem-header-height
       display: flex
