+++ src/components/navigation/problem-tab-content.sass
@@ -2,7 +2,7 @@
 
 .problem-tabs
   border-color: $problem-orange-light-3
-  border-width: 0 2px 2px 0
+  border-width: 0 $tab-section-border-width $tab-section-border-width 0
   border-style: solid
   .tab-list
     background-color: $problem-orange-light-7
