+++ src/components/thumbnail/tab-panel-documents-section.sass
@@ -7,7 +7,7 @@ $list-item-scale: 0.11
 .tab-panel-documents-section
   height: 100%
   .list
-    height: calc(100vh - (#{$header-height} + #{$workspace-content-margin} * 2 + #{$left-tab-height} * 2))
+    height: calc(100vh - (#{$header-height} + #{$workspace-content-margin} * 2 + #{$left-tab-height} * 2 + #{$tab-section-border-width} * 2))
     display: flex
     flex-wrap: wrap
     align-content: flex-start
