+++ src/components/vars.sass
@@ -316,6 +316,7 @@ $id-green: #91e7a2
 
 $header-height: 55px
 $left-tabs-width: 90px
+$tab-section-border-width: 2px
 $workspace-content-margin: 4px
 $problem-header-height: 38px
